get application runnable. figured out an issue with scraping: -1, 1, -1 tuples from the CSV API needed to have the middle 1 of (rank, level, xp) mapped to -1 in order to differentiate places where level is present data and actually is 1, vs the missing data case. this behavior is implemented in parse_stats_csv, so the CSV will be written that way on the next scrape, but the current raw scrape file would need doctoring to match this.
Former-commit-id: aa9bd39f433e2668dd2ca1e95660ee5006aa37a0

+++ src/analysis/results.py
@@ -7,10 +7,10 @@ from numpy.typing import NDArray
 def get_cluster_sizes(cluster_ids: NDArray) -> NDArray:
     """ Compute the number of occurrences for each cluster ID in an array. """
 
-    ids, counts = np.unique(cluster_ids.astype('int'), return_counts=True)
+    ids, counts = np.unique(cluster_ids, return_counts=True)
     sizes = np.zeros(max(ids) + 1)
     sizes[ids] = counts
-    return sizes
+    return sizes.astype('int')
 
 
 def get_cluster_uniqueness(cluster_sizes: NDArray) -> NDArray:
